Move MaterialXView's rendering codes out of Viewer class and define new abstraction call RenderPipeline that takes care of rendering. This change is done with goal of reducing code duplication.
